Fixed TADL-CMS sub-directory install
If TADL-CMS is installed in a sub-directory, it doesn't work properly. I fixed it by asking F3 for the URL instead of HTTP, since F3 is aware of the sub-directory we are installed in and will only return the relative path. 

Also, instead of removing the first 2 elements of the props, I only remove the first one, and the second ONLY if it is json. That way, in the future, we might be able to use the json controller even for non /json requests. Right now, they are NOT routed to this controller, but we are more future proof.